+++ src/model/model.ts
@@ -11,6 +11,8 @@ const model: Model = {
     return this.data;
   },
 
+  isOpenPopup: false,
+
   getProduct(productId) {
     const filterProducts = this.data.filter((product) => product.id === productId);
     const [product] = filterProducts;
