+++ src/types/model.ts
@@ -3,6 +3,7 @@ import { Products, Product } from './products';
 interface Model {
   data: Products;
   filteredData: Products;
+  isOpenPopup: boolean;
 
   getProducts(): Products;
   getProduct(productId: number): Product;
